refactor: name factory modules for the concept, not the create* fn
Match the existing convention (external-store.ts -> createExternalStore,
object-url.ts -> createObjectUrlRegistry): the file is named for the
concept, the factory keeps its create prefix.

- create-persisted-store.ts -> persisted-store.ts
- create-part-tracker.ts -> part-tracker.ts
